+++ lib/ts/nextjs.ts
@@ -12,7 +12,10 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
+import { serialize } from "cookie";
 import { errorHandler } from "./framework/express";
+import { CollectingResponse, PreParsedRequest, middleware } from "./framework/custom";
+import { HTTPMethod } from "./types";
 function next(
     request: any,
     response: any,
@@ -32,6 +35,21 @@ function next(
         });
     };
 }
+
+type PartialNextRequest = {
+    method: string;
+    url: string;
+    headers: Headers;
+    formData: () => any;
+    json: () => any;
+    cookies: {
+        getAll: () => {
+            name: string;
+            value: string;
+        }[];
+    };
+};
+
 export default class NextJS {
     static async superTokensNextWrapper<T>(
         middleware: (next: (middlewareError?: any) => void) => Promise<T>,
@@ -58,5 +76,57 @@ export default class NextJS {
             }
         });
     }
+
+    static getAppDirRequestHandler<T extends PartialNextRequest>(NextResponse: typeof Response) {
+        const stMiddleware = middleware<T>((req) => {
+            const query = Object.fromEntries(new URL(req.url).searchParams.entries());
+            const cookies: Record<string, string> = Object.fromEntries(
+                req.cookies.getAll().map((cookie) => [cookie.name, cookie.value])
+            );
+
+            return new PreParsedRequest({
+                method: req.method as HTTPMethod,
+                url: req.url,
+                query: query,
+                headers: req.headers,
+                cookies,
+                getFormBody: () => req.formData(),
+                getJSONBody: () => req.json(),
+            });
+        });
+
+        return async function handleCall(req: T) {
+            const baseResponse = new CollectingResponse();
+
+            const { handled, error } = await stMiddleware(req, baseResponse);
+
+            if (error) {
+                throw error;
+            }
+            if (!handled) {
+                return new NextResponse("Not found", { status: 404 });
+            }
+
+            for (const respCookie of baseResponse.cookies) {
+                baseResponse.headers.append(
+                    "Set-Cookie",
+                    serialize(respCookie.key, respCookie.value, {
+                        domain: respCookie.domain,
+                        expires: new Date(respCookie.expires),
+                        httpOnly: respCookie.httpOnly,
+                        path: respCookie.path,
+                        sameSite: respCookie.sameSite,
+                        secure: respCookie.secure,
+                    })
+                );
+            }
+
+            return new NextResponse(baseResponse.body, {
+                headers: baseResponse.headers,
+                status: baseResponse.statusCode,
+            });
+        };
+    }
 }
 export let superTokensNextWrapper = NextJS.superTokensNextWrapper;
+export let getAppDirRequestHandler = NextJS.getAppDirRequestHandler;