+++ lib/ts/utils.ts
@@ -6,9 +6,16 @@ import NormalisedURLPath from "./normalisedURLPath";
 import type { BaseRequest, BaseResponse } from "./framework";
 import { logDebugMessage } from "./logger";
 import { HEADER_FDI, HEADER_RID } from "./constants";
-import fetch from "cross-fetch";
+import crossFetch from "cross-fetch";
 import { User } from "./user";
 import { SessionContainer } from "./recipe/session";
+
+export const doFetch: typeof fetch = (...args) => {
+    if (typeof fetch !== "undefined") {
+        return fetch(...args);
+    }
+    return crossFetch(...args);
+};
 
 export function getLargestVersionFromIntersection(v1: string[], v2: string[]): string | undefined {
     let intersection = v1.filter((value) => v2.indexOf(value) !== -1);
@@ -299,7 +306,7 @@ export async function postWithFetch(
     let error;
     let resp: { status: number; body: any };
     try {
-        const fetchResp = await fetch(url, {
+        const fetchResp = await doFetch(url, {
             method: "POST",
             body: JSON.stringify(body),
             headers,