+++ recipe/session/framework/custom/index.d.ts
@@ -1,0 +1,3 @@
+export * from "../../../../lib/build/recipe/session/framework/custom";
+import * as _default from "../../../../lib/build/recipe/session/framework/custom";
+export default _default;